+++ src/window_type.h
@@ -732,7 +732,7 @@ enum GameOptionsInvalidationData {
 struct Window;
 
 /** Number to differentiate different windows of the same class */
-typedef int32_t WindowNumber;
+typedef int64_t WindowNumber;
 
 /** State of handling an event. */
 enum EventState {
